# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Text2SQL Studio API",
    "description": "Data Intelligence module \u2014 Evaluation Orchestration & Monitoring",
    "version": "2.0.0"
  },
  "paths": {
...
    "/feedback": {
      "post": {
        "tags": ["feedback"],
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_feedback_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/QueryFeedbackCreate" } }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/QueryFeedbackRead" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/feedback/table/{table_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Get Table Feedback",
        "operationId": "get_table_feedback_feedback_table__table_id__get",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Table Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/QueryFeedbackRead" },
                  "title": "Response Get Table Feedback Feedback Table  Table Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/feedback/query/{query_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Get Query Feedback",
        "operationId": "get_query_feedback_feedback_query__query_id__get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Query Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/QueryFeedbackRead" },
                  "title": "Response Get Query Feedback Feedback Query  Query Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackRating": {
        "type": "string",
        "enum": ["positive", "negative"],
        "title": "FeedbackRating"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryFeedbackCreate": {
        "properties": {
          "user_id": { "type": "string", "title": "User Id" },
          "query_id": { "type": "string", "title": "Query Id" },
          "table_id": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Table Id" },
          "rating": { "$ref": "#/components/schemas/FeedbackRating" },
          "comment": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Comment" },
          "suggested_correction": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Suggested Correction"
          }
        },
        "type": "object",
        "required": ["user_id", "query_id", "rating"],
        "title": "QueryFeedbackCreate"
      },
      "QueryFeedbackRead": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "user_id": { "type": "string", "title": "User Id" },
          "query_id": { "type": "string", "title": "Query Id" },
          "table_id": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Table Id" },
          "rating": { "$ref": "#/components/schemas/FeedbackRating" },
          "comment": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Comment" },
          "suggested_correction": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Suggested Correction"
          },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "query_id",
          "table_id",
          "rating",
          "comment",
          "suggested_correction",
          "created_at"
        ],
        "title": "QueryFeedbackRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}